requirements.txt
fastapi
uvicorn
streamlit
requests
pydantic

openai
python-dotenv
